uint32 to uint64

+++ vowpalwabbit/core/include/vw/core/array_parameters.h
@@ -39,7 +39,7 @@ public:
   using pointer = T*;
   using reference = T&;
 
-  sparse_iterator(weight_map::iterator& iter, uint32_t stride) : _iter(iter), _stride(stride) {}
+  sparse_iterator(weight_map::iterator& iter, uint64_t stride) : _iter(iter), _stride(stride) {}
 
   sparse_iterator& operator=(const sparse_iterator& other) = default;
   sparse_iterator(const sparse_iterator& other) = default;
@@ -177,7 +177,7 @@ public:
 
   uint64_t seeded() const { return _seeded; }
 
-  uint32_t stride() const { return 1 << _stride_shift; }
+  uint64_t stride() const { return static_cast<uint64_t>(1) << _stride_shift; }
 
   uint32_t stride_shift() const { return _stride_shift; }
 
@@ -240,7 +240,7 @@ public:
     }
   }
 
-  inline uint32_t stride() const
+  inline uint64_t stride() const
   {
     if (sparse) { return sparse_weights.stride(); }
     else
